requirements.txt
fastapi
uvicorn
python-multipart
pydub
sqlalchemy
psycopg2
pydantic
python-jose
passlib
bcrypt
pytube
scipy
torch
torchaudio
transformers
datasets
accelerate
sentencepiece
